% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbcm.di.R
\name{pbcm.di}
\alias{pbcm.di}
\title{Data-informed Parametric Bootstrap Cross-fitting}
\usage{
pbcm.di(data, fun1, fun2, genfun1, genfun2, reps, args1 = NULL,
  args2 = NULL, genargs1 = NULL, genargs2 = NULL,
  print_genargs = TRUE, nonparametric_bootstrap = TRUE,
  verbose = TRUE, GoFname = "GoF")
}
\arguments{
\item{data}{Data frame}

\item{fun1}{First modelling function}

\item{fun2}{Second modelling function}

\item{genfun1}{Generator function for first model}

\item{genfun2}{Generator function for second model}

\item{reps}{Number of Monte Carlo repetitions}

\item{args1}{List of arguments passed to \code{fun1}}

\item{args2}{List of arguments passed to \code{fun2}}

\item{genargs1}{List of arguments passed to \code{genfun1}}

\item{genargs2}{List of arguments passed to \code{genfun2}}

\item{print_genargs}{Whether the generator argument values should be included in output (see Details)}

\item{nonparametric_bootstrap}{Whether \code{data} should be nonparametrically bootstrapped before the parametric bootstrap}

\item{verbose}{If \code{TRUE}, a progress bar is printed to the console and warnings are issued}

\item{GoFname}{Name of the element returned by \code{fun1} and \code{fun2} holding the goodness of fit; see Details}
}
\value{
A data frame in long format with the following columns:
\describe{
\item{\code{rep}}{Monte Carlo repetition number}
\item{\code{generator}}{Generating model}
\item{\code{GoF1}}{Goodness of fit of model 1}
\item{\code{GoF2}}{Goodness of fit of model 2}
\item{\code{DeltaGoF}}{Equals \code{GoF1 - GoF2}}
}
In addition to these columns, if \code{print_genargs == TRUE}, each argument in the lists \code{genargs1} and \code{genargs2} is included as a column of its own, with the argument's name prefixed by \code{"genargs1_"} or \code{"genargs2_"}.
}
\description{
The data-informed Parametric Bootstrap Cross-fitting Method (PBCM) generates synthetic data from two models of a phenomenon parameterized by fits to an empirical dataset, and then cross-fits the models to these data. The result is two distributions of the goodness of fit metric \eqn{\Delta GoF = GoF_1 - GoF_2}, where \eqn{GoF_1} is the fit of model 1 and \eqn{GoF_2} the fit of model 2.
}
\details{
Functions \code{fun1} and \code{fun2} must take \code{data} as an argument in addition to any arguments specified in \code{args1} and \code{args2}. Moreover, these functions must return a list with at least one element carrying the goodness of fit; the name of this element may be specified through the \code{GoFname} argument, by default the string \code{"GoF"} is assumed. Functions \code{genfun1} and \code{genfun2} must take an argument named \code{model} (the output of \code{fun1} and \code{fun2}).
}
\examples{
x <- seq(from=0, to=1, length.out=100)
mockdata <- data.frame(x=x, y=x + rnorm(100, 0, 0.5))

myfitfun <- function(data, p) {
  res <- nls(y~a*x^p, data, start=list(a=1.1))
  list(a=coef(res), GoF=deviance(res))
}

mygenfun <- function(model, p) { 
  x <- seq(from=0, to=1, length.out=100)
  y <- model$a*x^p + rnorm(100, 0, 0.5)
  data.frame(x=x, y=y)
}

pbcm.di(data=mockdata, fun1=myfitfun, fun2=myfitfun, genfun1=mygenfun,
        genfun2=mygenfun, reps=20, args1=list(p=1), args2=list(p=2), 
        genargs1=list(p=1), genargs2=list(p=2))
}
\references{
Wagenmakers, E.-J., Ratcliff, R., Gomez, P. & Iverson, G. J. (2004) Assessing model mimicry using the parametric bootstrap. \emph{Journal of Mathematical Psychology}, 48(1), 28–50. \url{https://doi.org/10.1016/j.jmp.2003.11.004}
}
\author{
Henri Kauhanen
}
