% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbcm.du.R
\name{pbcm.du}
\alias{pbcm.du}
\title{Data-uninformed Parametric Bootstrap Cross-fitting}
\usage{
pbcm.du(
  fun1,
  fun2,
  genfun1,
  genfun2,
  genargs1,
  genargs2,
  reps,
  GOF = "RSS",
  progressbar = TRUE
)
}
\arguments{
\item{fun1}{First modelling function}

\item{fun2}{Second modelling function}

\item{genfun1}{Generator function for first model}

\item{genfun2}{Generator function for second model}

\item{genargs1}{List of arguments passed to first generator function}

\item{genargs2}{List of arguments passed to second generator function}

\item{reps}{Number of Monte Carlo repetitions}

\item{GOF}{FIXME}

\item{progressbar}{Whether a (text) progress bar ought to be printed}
}
\value{
FIXME
}
\description{
description
}
\details{
details
}
