% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbcm.du.R
\name{pbcm.du}
\alias{pbcm.du}
\title{Data-uninformed Parametric Bootstrap Cross-fitting}
\usage{
pbcm.du(fun1, fun2, genfun1, genfun2, reps, args1 = NULL, args2 = NULL,
  genargs1 = NULL, genargs2 = NULL, verbose = TRUE,
  GoFname = "GoF")
}
\arguments{
\item{fun1}{First modelling function}

\item{fun2}{Second modelling function}

\item{genfun1}{Generator function for first model}

\item{genfun2}{Generator function for second model}

\item{reps}{Number of Monte Carlo repetitions}

\item{args1}{List of arguments passed to \code{fun1}}

\item{args2}{List of arguments passed to \code{fun2}}

\item{genargs1}{List of arguments passed to \code{genfun1}}

\item{genargs2}{List of arguments passed to \code{genfun2}}

\item{verbose}{If \code{TRUE}, a progress bar is printed to the console and warnings are issued}

\item{GoFname}{Name of the element returned by \code{fun1} and \code{fun2} holding the goodness of fit; see Details}
}
\value{
A data frame in long format with the following columns:
\describe{
\item{\code{rep}}{Monte Carlo repetition number}
\item{\code{generator}}{Generating model}
\item{\code{GoF1}}{Goodness of fit of model 1}
\item{\code{GoF2}}{Goodness of fit of model 2}
\item{\code{DeltaGoF}}{Equals GoF1 - GoF2}
}
In addition to these columns, each argument in the lists \code{genargs1} and \code{genargs2} is included as a column of its own, with the argument's name prefixed by "model1_" or "model2_".
}
\description{
The data-uninformed Parametric Bootstrap Cross-fitting Method (PBCM) generates synthetic data from two models of a phenomenon with given model parameter values, and then cross-fits the models to these data. The result is two distributions of the goodness of fit metric DeltaGoF = GoF1 - GoF2, where GoF1 is the fit of model 1 and GoF2 the fit of model 2.
}
\details{
Functions \code{fun1} and \code{fun2} must take an argument named \code{data} in addition to any arguments specified in \code{args1} and \code{args2}; this is used to pass the synthetic data generated by \code{genfun1} and \code{genfun2}. \code{fun1} and \code{fun2} must return a list with at least one element carrying the goodness of fit; the name of this element may be specified through the \code{GoFname} argument, by default "GoF" is assumed.
}
\references{
Wagenmakers, E.-J., Ratcliff, R., Gomez, P. & Iverson, G. J. (2004) Assessing model mimicry using the parametric bootstrap. \emph{Journal of Mathematical Psychology}, 48(1), 28-50.
}
\author{
Henri Kauhanen
}
